material and category
make universal units and make buy material popup

+++ database/migrations/2026_01_27_061036_create_materials_table.php
@@ -15,10 +15,9 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->foreignId('unit_id')->constrained('units')->onDelete('restrict');
-            $table->integer('qty');
+            $table->float('qty');
             $table->timestamps();
             $table->softDeletes();
-
         });
     }
